requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
python-dotenv==1.0.1
groq>=0.9.0
httpx==0.27.0
requests==2.32.3
gspread==6.1.2
google-auth==2.30.0
google-auth-oauthlib==1.2.1
apscheduler==3.10.4
beautifulsoup4==4.12.3
pdfplumber==0.11.1
pandas>=2.2.2
openpyxl==3.1.4
pillow==10.3.0
pytesseract==0.3.10
python-multipart==0.0.9
jinja2==3.1.4
aiofiles==23.2.1
pytz==2024.1
websockets
pydub